ContentSize for ScrollView on IB
The StackView creates a strange link between the framesize and the
contentsize. As the StackView is adaptable, then both the framesize and
the contentsize will be the same due to constraints. To break that, we
attach the width to a third element (the arrow keys), and that makes the
contentsize separate from the frame, which then will be controlled as
normal by the Stack. PITA